% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Polynomial_Group_structure.R
\name{MEM_Polynomial_Group_structure}
\alias{MEM_Polynomial_Group_structure}
\title{Polynomial Mixed-Effects Models with Censored and Group-Structured Responses}
\usage{
MEM_Polynomial_Group_structure(
  y,x = NULL,Group = NULL,Id = NULL,Cens = NULL,
  marginal_dyn_type = "polynomial",ind_dyn_type = "polynomial",
  global_intercept = TRUE,group_intercept = FALSE,
  degree_group = 3,Adaptive = "none",knotnumcrit = "AIC",
  min_knots_group = 2,max_knots_group = 2,
  knots_group = NULL,df_group = NULL,Boundary.knots_group = NULL,
  ind_intercept = FALSE,degree_ind = 3,
  min_knots_ind = 2,max_knots_ind = 2,
  same_base_group_ind = FALSE,knots_ind = NULL,
  df_ind = NULL,Boundary.knots_ind = NULL,
  ...
)
}
\arguments{
\item{y}{observed responses described either as a data frame containing at least a column named \emph{'y'} and possibly the columns \emph{'x'}, \emph{'Group'}, \emph{'Id'} and \emph{'Cens'} (among others), or as a vector of numerical values.}

\item{x}{a numerical vector representing the x-axis coordinates corresponding to the observed responses (e.g. time of observations) which can be defined is \code{y} is a vector or a data frame without \code{x} column. By default, this variable is defined as NULL.}

\item{Group}{a vector of group indicator for each observed responses which can be defined if \code{y} is a vector or a data frame without \code{Group} column. If this variable is defined as NULL (default) and \emph{y} does not contain group information, all observed data are assumed to belong to the same group.}

\item{Id}{a vector of individual ID for each observed responses which can be defined if \code{y} is a vector or a data frame without \emph{'Id'} column. By default, this variable is defined as NULL.}

\item{Cens}{a vector of censoring indicator (if \code{y} >= ytrue, then \code{Cens} == 1). If this variable is defined as NULL (default) and \code{y} does not contain \emph{'Cens'} column, observed data are assumed as uncensored.}

\item{marginal_dyn_type}{a character variable indicating the type of marginal dynamics. Options are 'polynomial' (default) and 'spline'.}

\item{ind_dyn_type}{a character variable indicating the type of individual dynamics (random effects). Options are 'polynomial' (default) or 'spline'.}

\item{global_intercept}{a logical scalar. If TRUE (default) a global intercept (no group-specific) is included in the marginal dynamics.}

\item{group_intercept}{a logical scalar (same option for all groups) or vector. For each group, if TRUE, a group-specific intercept is included in the marginal dynamics. By default, this variable is defined as FALSE}

\item{degree_group}{an integer scalar (same option for all groups) or vector. The variable indicates for each group either the degree of polynomial functions or spline curves describing marginal dynamics. By default, the variable is fixed at 3.}

\item{Adaptive}{an optional character variable whether B-spline curves are built with internal knot positions optimally estimated according to data (see \link[AUCcomparison]{Optimal_knot_research} and @details for more details). Options are 'none' (default), 'group , 'individual', and 'both'.}

\item{knotnumcrit}{an optional character variable indicating the criterion to be used for determining the number of internal knots in B-splines. See @details for more details.}

\item{min_knots_group}{an optional integer scalar indicating the minimum number of internal knots to consider in the research of optimal knots for marginal dynamics. By default, this variable fixed at 2. See @details for more details.}

\item{max_knots_group}{an optional integer scalar indicating the maximum number of internal knots to consider in the research of optimal knots for marginal dynamics. By default, this variable fixed at 2. See @details for more details.}

\item{knots_group}{a numerical vector (same option for all groups) or a list of either numerical vectors or NULL (one for each group) indicating the internal knots for group-specific B-spline curves. By default, this variable is defined as NULL (see \link[splines]{bs} and @details for more details).}

\item{df_group}{an integer scalar (same option for all groups) or vector indicating the degrees of freedom to consider to build marginal B-spline curves. One can specify \code{df_group} rather than \code{knots_group} (see \link[splines]{bs} for more details). By default, this variable is defined as NULL. See @details for more details.}

\item{Boundary.knots_group}{a numerical vector (same option for all groups) or a list of either numerical vectors or NULL (one for each group) indicating the boundary knots for group-specific B-spline curves. By default, this variable is defined as NULL (see \link[splines]{bs} and @details for more details).}

\item{ind_intercept}{a logical scalar. If TRUE, an intercept is included in the individual dynamics (random effects). By default, this variable is defined as FALSE.}

\item{degree_ind}{an integer scalar indicating either the degree of the polynomial functions or the B-spline curves describing individual dynamics. By default, this variable is fixed at 2.}

\item{min_knots_ind}{an optional interger scalar indicating the minimum number of internal knots to consider in the research of optimal knots for individual dynamics. By default, this variable is fixed at 2. See @details for more details.}

\item{max_knots_ind}{an optional interger scalar indicating the maximum number of internal knots to consider in the research of optimal knots for individual dynamics. By default, this variable is fixed at 2. See @details for more details.}

\item{same_base_group_ind}{an optional logical scalar indicating whether or not the same B-spline basis must be considered in group-specific and individual dynamics. If TRUE, each individual B-spline basis will be build as the corresponding group-specific B-spline basis evaluated at the individual predictor variable. By default, this variable is defined as FALSE. See @details for more details.}

\item{knots_ind}{a numerical vector (same option for all individuals) or a list of numerical values or NULL indicating the internal knots for individual-specific B-spline curves. If this variable is defined as a list, internal knots can either be defined individually (one vector or NULL for each Id value), equivalent for each individual belong to the same group (one vector or NULL for each Group), or equivalent for each individual (one vector or NULL). By default, this variable is defined as NULL (see \link[splines]{bs} and @details for more details).}

\item{df_ind}{an integer scalar (same option for all individuals) or vector indicating the degrees of freedom to consier to build individual B-splines curves. This variable can be choosen different for each individual or equivalent for each individual belonging to the same group (one value for each group). By default, this variable is defined as NULL(see \link[splines]{bs} and @details for more details).}

\item{Boundary.knots_ind}{a numerical vector indicating the boundary knots for individual-specific B-spline curves. By default, this variable is defined as NULL (see \link[splines]{bs} and @details for more details).}

\item{...}{Further arguments to be passed (see \link[lmec]{lmec} for more details).}
}
\value{
A list containing:
\itemize{
\item \code{Model_estimation}:  a list containing the results of the model estimation provided by the function \link[lmec]{lmec}. In this list, the vector of fixed parameters called \code{beta}, the parameters are returned in the following order: \mjteqn{\beta = (\gamma_0,\beta_0^1, \cdots, \beta_{K_1}^1,\cdots,\beta_0^g,\cdots,\beta_{K_g}^g,\cdots,\beta_0^G,\cdots,\beta_{K_G}^G)}{\beta = (\gamma_0,\beta_0^1, \cdots, \beta_{K_1}^1,\cdots,\beta_0^g,\cdots,\beta_{K_g}^g,\cdots,\beta_0^G,\cdots,\beta_{K_G}^G)}{\beta = (\gamma_0,\beta_0^1, \cdots, \beta_{K_1}^1,\cdots,\beta_0^g,\cdots,\beta_{K_g}^g,\cdots,\beta_0^G,\cdots,\beta_{K_G}^G)}.
\item \code{Model_features}:  a list of 3 elements:
\enumerate{
\item \code{Groups}:   a vector indicating the names of the groups involved in the model.
\item \code{Marginal.dyn.feature}:  a list summarizing the features of the marginal dynamics defined in the model (through input arguments):
\itemize{
\item \code{dynamic.type}: a character scalar indicating the chosen type of marginal dynamics.
\item \code{intercept}: a logical vector summarizing choices about global and group-specific intercepts (Number of groups + 1 values).
}
For 'polynomial' marginal dynamics: \itemize{
\item \code{polynomial.degree}: an integer vector indicating the degree of polynomial functions.
}
For 'spline' marginal dynamics: \itemize{
\item \code{spline.degree}: an integer vector indicating the degree of B-spline curves.
\item \code{adaptive.splines}: a logical scalar indicating whether or not adaptive internal knots have been considered.
\item \code{knots}: a list of group-specific internal knots used to build B-spline basis. If the degrees of freedom were equals to the spline degrees, then NULL.
\item \code{df}: a numerical vector of group-specific degrees of freedom used to build B-spline basis.
\item \code{boundary.knots}: a list of group-specific boundary knots used to build B-spline  basis.
}
\item \code{Individual.dyn.feature}: a list summarizing the features of the individual dynamics defined in the model (through input arguments) \itemize{
\item \code{dynamic.type}: a character scalar indicating the chosen type of individual dynamics.
\item \code{intercept}: a logical scalar indicating whether a random intercept was included in the model.
}
For 'polynomial' individual dynamics: \itemize{
\item \code{polynomial.degree}: an integer scalar indicating the degree of polynomial functions.
}
For 'spline' marginal dynamics: \itemize{
\item \code{spline.degree}: an integer scalar indicating the degree of B-spline curves
\item \code{adaptive.splines}: a logical scalar indicating whether or not adaptive internal knots have been considered.
\item \code{knots}: a data frame of individually estimated internal knots (if \code{Adaptive} chosen as 'individual' or 'both'), or a list of chosen individual internal knots.
\item \code{df}: a numerical vector of individual degrees of freedom.
\item \code{boundary.knots}: a numerical vector of individual boundary knots.
}
}
}
}
\description{
\loadmathjax This function fits a mixed-effects model (MEM) to potentially censored data structured by group when marginal and individual dynamics are described either by polynomials or B-spline curves.
}
\details{
The variable \code{adaptive} can take 4 differents values: \itemize{
\item 'none' (default): internal knot positions defining B-spline curves in the model are not optimally estimated according to data, whether for fixed or random effects. Knots are then defined by users.
\item 'group': only internal knot positions defining B-spline curves in fixed effects are optimally estimated, even if random effects involve B-splines.
\item 'individual': only internal knot positions defining B-spline curves in random effects are  optimally estimated, even if fixed effects involve B-splines.
\item 'both': internal knot positions defining B-spline curves for both fixed and random effects are optimally estimated.
}

At group level (fixed effects), \itemize{
\item the variables \code{Adaptive}, \code{knotnumcrit}, \code{min_knots_group}, \code{max_knots_group}, \code{knots_group}, \code{df_group} and \code{Boundary.knots_group} will be used only if the variable \code{marginal_dyn_type} has been chosen as 'spline'.
\itemize{
\item among them, the variables \code{knotnumcrit},  \code{min_knots_group} and \code{max_knots_group} will be used only if adaptive knots are requested (the variable \code{Adaptive} chosen as 'group' or 'both').
\item the variables \code{knots_group}, \code{df_group} and \code{Boundary.knots_group} will be used without adaptive knots.
}
}
Similarly, at individual level (random effects), \itemize{
\item the variables \code{Adaptive}, \code{knotnumcrit}, \code{min_knots_ind}, \code{max_knots_ind}, \code{knots_ind}, \code{df_ind} and \code{Boundary.knots_ind} will be used only if the variable \code{ind_dyn_type} has been chosen as 'spline'.
\itemize{
\item among them, the variables \code{knotnumcrit}, \code{min_knots_ind} and \code{max_knots_ind} will be used only if adaptive knots are requested (the variable \code{Adaptive} chosen as 'individual' or 'both').
\item the variables \code{knots_ind}, \code{df_ind} and \code{Boundary.knots_ind} will be used without adaptive knots.
}
\item the variable \code{same_base_group_ind} can be used only if both variables \code{marginal_dyn_type} and \code{ind_dyn_type} are chosen as 'spline'.
}

The Mixed-Effects model describing the outcome of interest \mjteqn{Y_{ij,g_i}}{Y_{ij,g_i}}{Y_{ij,g_i}} of the subject \mjteqn{i}{i}{i} in the group \mjteqn{g_i}{g_i}{g_i} at the \mjteqn{j}{j}{j}th time point (x-axis) is given by
\mjtdeqn{Y_{ij,g_i} = \gamma_0 +  \sum_{g=1}^{G} 1_{g_i=g}\times F_g(t_{ij,g}) + h_i(t_{ij,g_i}) + \varepsilon_{ij} }{Y_{ij,g_i} = \gamma_0 +  \sum_{g=1}^{G} 1_{g_i=g}\times F_g(t_{ij,g}) + h_i(t_{ij,g_i}) + \varepsilon_{ij} }{Y_{ij,g_i} = \gamma_0 +  \sum_{g=1}^{G} 1_{g_i=g} \times F_g(t_{ij,g}) + h_i(t_{ij,g_i}) + \varepsilon_{ij}}
where \mjseqn{\gamma_0} is the global (non group-specific) intercept and the functions \mjteqn{F_g}{F_g}{F_g} and \mjteqn{h_i}{h_i}{h_i} are the non-linear smooth functions describing respectively the marginal group-specific dynamics and the individual dynamics (random effects).
Through this function, the group-specific functions are defined as following, where \mjteqn{\beta_0^g}{\beta_0^g}{\beta_0^g} is the optional \code{group_intercept}:
\mjtdeqn{F_g(t_{ij,g}) = \beta_0^g + \sum_{k=1}^{K_g} \beta^g_k f^k_g(t_{ij,g})}{F_g(t_{ij,g}) = \beta_0^g + \sum_{k=1}^{K_g} \beta^g_k f^k_g(t_{ij,g})}{F_g(t_{ij,g}) = \beta_0^g + \sum_{k=1}^{K_g} \beta^g_k f^k_g(t_{ij,g})}
If \code{marginal_dyn_type} is defined as 'polynomial', \mjteqn{K_g}{K_g}{K_g} = \code{degree_group} and \mjteqn{f^k_g(t_{ij,g}) = t_{ij,g}^k}{f^k_g(t_{ij,g}) = t_{ij,g}^k}{f^k_g(t_{ij,g}) = t_{ij,g}^k} and if \code{marginal_dyn_type} is defined as 'spline', \mjteqn{K_g}{K_g}{K_g} = \code{df_group} (see \link[splines]{bs} for more details) and \mjteqn{f^k_g(t_{ij,g})}{f^k_g(t_{ij,g})}{f^k_g(t_{ij,g})} is the \mjteqn{k}{k}{k}th group-specific spline basis.
Similarly, the individual dynamics are described by the following functions, with \mjteqn{b_{0i}}{b_{0i}}{b_{0i}} as optional \code{ind_intercept}:
\mjtdeqn{h_i(t_{ij,g}) = b_{0i} + \sum_{k=1}^{K_i} b_{ki} \Psi_k^i(t_{ij,g})}{h_i(t_{ij,g}) = b_{0i} + \sum_{k=1}^{K_i} b_{ki} \Psi_k^i(t_{ij,g})}{h_i(t_{ij,g}) = b_{0i} + \sum_{k=1}^{K_i} b_{ki} \Psi_k^i(t_{ij,g})}
If \code{ind_dyn_type} is defined as 'polynomial', \mjteqn{K_i}{K_i}{K_i} = \code{degree_ind} and \mjteqn{\Psi_k^i(t_{ij,g}) = t_{ij,g}^k}{\Psi_k^i(t_{ij,g}) = t_{ij,g}^k}{\Psi_k^i(t_{ij,g}) = t_{ij,g}^k} and if \code{ind_dyn_type} is defined as 'spline', \mjteqn{K_i}{K_i}{K_i} = \code{df_ind} (see \link[splines]{bs} for more details) and \mjteqn{\Psi_k^i(t_{ij,g})}{\Psi_k^i(t_{ij,g})}{\Psi_k^i(t_{ij,g})} is the  \mjteqn{k}{k}{k}th individual spline basis.
}
\examples{
\donttest{ # Download of data
data("HIV_Simu_Dataset_Delta01_cens")
data <- HIV_Simu_Dataset_Delta01_cens

# Change factors in character vectors
data$id <- as.character(data$id) ; data$Group <- as.character(data$Group)

# We call the function considering the variable 'y' as a vector (we need to specify the groups )
MEM_Pol_Group <- MEM_Polynomial_Group_structure(y=data$VL,x=data$time,Group=data$Group,
                                                Id=data$id,Cens=data$cens)

# We call the function considering the variable 'y' as a  data frame
colnames(data)[4] <- "y"
MEM_Pol_Group <- MEM_Polynomial_Group_structure(y=data,x=data$time,Cens=data$cens,Id=data$id)} 


}
\seealso{
\code{\link[splines]{bs}},
\code{\link[lmec]{lmec}},
\code{\link[AUCcomparison]{Optimal_knot_research}}
}
